services/transcription/requirements.txt
# Core FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Audio processing
ffmpeg-python==0.2.0
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
webrtcvad==2.0.10
noisereduce==3.0.0

# OpenAI and ML
openai==1.3.5
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
datasets==2.14.6
accelerate==0.24.1
numpy==1.24.3
scipy==1.11.4

# Redis and caching
redis[hiredis]==5.0.1
aioredis==2.0.1

# HTTP and networking
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Data processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-json-logger==2.0.7

# Medical and language processing
spacy==3.7.2
nltk==3.8.1
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0

# Google AI
google-generativeai==0.3.2

# Monitoring and metrics
prometheus-client==0.19.0
psutil==5.9.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
click==8.1.7
tenacity==8.2.3
structlog==23.2.0